applied fixes from M. Laukkanen
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2761 06827809-a52a-0410-b366-d66718629ded

+++ apps/encode_keychange.c
@@ -111,7 +111,7 @@ int
 main(int argc, char **argv)
 {
 	int		  rval		= SNMPERR_SUCCESS;
-	u_int		  oldKu_len	= SNMP_MAXBUF_SMALL,
+	size_t		  oldKu_len	= SNMP_MAXBUF_SMALL,
 			  newKu_len	= SNMP_MAXBUF_SMALL,
 			  oldkul_len	= SNMP_MAXBUF_SMALL,
 			  newkul_len	= SNMP_MAXBUF_SMALL,
